requirements.txt
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
pillow>=10.2.0
tqdm>=4.66.0
streamlit>=1.32.0
matplotlib>=3.8.0
numpy>=1.26.0
